Move the password change request
