+++ src/Microsoft.Data.SqlClient/netcore/src/Microsoft.Data.SqlClient.csproj
@@ -919,7 +919,7 @@
       <DependentUpon>Strings.resx</DependentUpon>
     </Compile>
     <Compile Include="Microsoft\Data\SqlClientX\Handlers\Connection\LoginHandler.cs" />
-    <Compile Include="Microsoft\Data\SqlClientX\Handlers\Connection\Login\PasswordChangeRequest.cs" />
+    <Compile Include="Microsoft\Data\SqlClientX\Handlers\PasswordChangeRequest.cs" />
     <Compile Include="Microsoft\Data\SqlClientX\Handlers\Connection\Login\TdsFeatures.cs" />
     <Compile Include="Microsoft\Data\SqlClientX\Handlers\Connection\PreloginHandler.cs" />
     <Compile Include="Microsoft\Data\SqlClientX\Handlers\ConnectionHandlerContext.cs" />
